gen_lib tweaks and cleanup; try to improve how the parameter space is handled to remove dependence on 'Parameter_Space' class explicitly
